moves firebase endpoints to server admin files, adds front end to interact, updates get_firebase_instance to check rather than strictly update app, adds misc logic to use get_firebase_instance instead of old config variable
