+++ frontend/templates/manage_firebase_credentials.html
@@ -7,9 +7,11 @@
 
       <!-- firebase cert upload -->
       {% if not firebase_credentials_exists %}
+
         <h2>Upload a Firebase Notification Certificate</h2>
+        <div class="text-danger"><h4>No Existing Firebase Certificate</h4></div>
+
         <form action="/upload_backend_firebase_cert" method="post" class="form-horizontal" enctype="multipart/form-data">
-        <!-- Submit button -->
           <div class="form-group">
             <div class="col-xs-offset-3 col-xs-6">
               <input type="file" name="backend_firebase_cert" id="backend_firebase_cert">
@@ -23,11 +25,43 @@
       {% else %}
       <br><br><hr>
         <h2>Delete or Overwrite the Existing Firebase Notification Certificate</h2>
-        <form action="/upload_backend_firebase_cert" method="post" class="form-horizontal" enctype="multipart/form-data">
-        <!-- Submit button -->
+        <div class="text-success"><h4>A Firebase certificate has already been uploaded</h4></div>
+        <br>
+        <div class="row">
+          <div class="col-xs-offset-2 col-xs-4">
+            <form action="/upload_backend_firebase_cert" method="post" class="form-horizontal" enctype="multipart/form-data">
+              <div class="form-group">
+                  <input type="file" name="backend_firebase_cert" id="backend_firebase_cert">
+                <br><br>
+                <div class="col-xs-4">
+                  <button type="submit" class="btn btn-primary">Overwrite</button>
+                </div>
+              </div>
+            </form>
+          </div>
+          <div class="col-xs-4">
+            <form action="/delete_backend_firebase_cert" method="post" class="form-horizontal">
+              <div class="col-xs-offset-3 col-xs-6">
+                  <button type="submit" class="btn btn-danger">Delete</button>
+                </div>
+            </form>
+          </div>
+        </div>
+
+      {% endif %}
+      <br><br>
+
+    <!-- android cert upload -->
+    <br><br><hr>
+    {% if not android_credentials_exists %}
+
+        <h2>Upload an Android Firebase services file</h2>
+        <div class="text-danger"><h4>No Credentials</h4></div>
+
+        <form action="/upload_android_firebase_cert" method="post" class="form-horizontal" enctype="multipart/form-data">
           <div class="form-group">
             <div class="col-xs-offset-3 col-xs-6">
-              <input type="file" name="backend_firebase_cert" id="backend_firebase_cert">
+              <input type="file" name="android_firebase_cert" id="android_firebase_cert">
             </div>
             <br><br>
             <div class="col-xs-offset-3 col-xs-6">
@@ -35,13 +69,82 @@
             </div>
           </div>
         </form>
-        <form action="/delete_backend_firebase_cert" method="post" class="form-horizontal">
-          <div class="col-xs-offset-3 col-xs-6">
-              <button type="submit" class="btn btn-danger">Delete</button>
-            </div>
-        </form>
+      {% else %}
+        <h2>Delete or Overwrite the Existing Android Firebase services file</h2>
+        <div class="text-success"><h4>An Android Firebase services file has already been uploaded</h4></div>
+        <br>
+        <div class="row">
+          <div class="col-xs-offset-2 col-xs-4">
+            <form action="/upload_android_firebase_cert" method="post" class="form-horizontal" enctype="multipart/form-data">
+              <div class="form-group">
+                  <input type="file" name="upload_android_firebase_cert" id="android_firebase_cert">
+                <br><br>
+                <div class="col-xs-4">
+                  <button type="submit" class="btn btn-primary">Overwrite</button>
+                </div>
+              </div>
+            </form>
+          </div>
+          <div class="col-xs-4">
+            <form action="/delete_android_firebase_cert" method="post" class="form-horizontal">
+              <div class="col-xs-offset-3 col-xs-6">
+                  <button type="submit" class="btn btn-danger">Delete</button>
+                </div>
+            </form>
+          </div>
+        </div>
+
       {% endif %}
+
       <br><br>
+
+        <!-- ios cert upload -->
+    <br><br><hr>
+    {% if not ios_credentials_exists %}
+
+        <h2>Upload an IOS Firebase plist file</h2>
+        <div class="text-danger"><h4>No Credentials</h4></div>
+
+        <form action="/upload_ios_firebase_cert" method="post" class="form-horizontal" enctype="multipart/form-data">
+          <div class="form-group">
+            <div class="col-xs-offset-3 col-xs-6">
+              <input type="file" name="ios_firebase_cert" id="ios_firebase_cert">
+            </div>
+            <br><br>
+            <div class="col-xs-offset-3 col-xs-6">
+              <button type="submit" class="btn btn-primary">Upload</button>
+            </div>
+          </div>
+        </form>
+      {% else %}
+        <h2>Delete or Overwrite the Existing IOS Firebase Notification Certificate</h2>
+        <div class="text-success"><h4>An IOS Firebase plist file has already been uploaded</h4></div>
+        <br>
+        <div class="row">
+          <div class="col-xs-offset-2 col-xs-4">
+            <form action="/upload_ios_firebase_cert" method="post" class="form-horizontal" enctype="multipart/form-data">
+              <div class="form-group">
+                  <input type="file" name="ios_firebase_cert" id="ios_firebase_cert">
+                <br><br>
+                <div class="col-xs-4">
+                  <button type="submit" class="btn btn-primary">Overwrite</button>
+                </div>
+              </div>
+            </form>
+          </div>
+          <div class="col-xs-4">
+            <form action="/delete_ios_firebase_cert" method="post" class="form-horizontal">
+              <div class="col-xs-offset-3 col-xs-6">
+                  <button type="submit" class="btn btn-danger">Delete</button>
+                </div>
+            </form>
+          </div>
+        </div>
+
+      {% endif %}
+
+      <br><br>
+
 
     </div>
