+++ libs/push_notifications.py
@@ -6,7 +6,7 @@ from django.utils.timezone import is_aware, is_naive, make_aware
 from firebase_admin import credentials, initialize_app as initialize_firebase_app, \
     delete_app as delete_firebase_instance, get_app as get_firebase_app
 
-from config.constants import BACKEND_FIREBASE_CREDENTIALS
+from config.constants import ANDROID_FIREBASE_CREDENTIALS, BACKEND_FIREBASE_CREDENTIALS, IOS_FIREBASE_CREDENTIALS
 from config.settings import PUSH_NOTIFICATIONS_ENABLED
 from database.schedule_models import (AbsoluteSchedule, ArchivedEvent, ScheduledEvent,
     WeeklySchedule)
@@ -15,22 +15,24 @@ from database.system_models import FileAsText
 from database.user_models import Participant
 
 
-def update_firebase_instance():
+def get_firebase_instance(credentials_updated=False, require_android=False, require_ios=False):
     """ Ensure that the current firebase credentials being used reflect the state of the database, including possibly
      removing the app if credentials have been removed. This function should be called after any update to the stored
      credentials and on startup """
-    try:
-        delete_firebase_instance(get_firebase_app())
-    except ValueError:
-        pass
-    if not FileAsText.objects.filter(tag=BACKEND_FIREBASE_CREDENTIALS).exists():
+    if credentials_updated:
+        try:
+            delete_firebase_instance(get_firebase_app())
+        except ValueError:
+            pass
+        if not FileAsText.objects.filter(tag=BACKEND_FIREBASE_CREDENTIALS).exists():
+            return None
+        stored_credentials = json.loads(FileAsText.objects.get(tag=BACKEND_FIREBASE_CREDENTIALS).text)
+        return initialize_firebase_app(credentials.Certificate(stored_credentials))
+    if (not FileAsText.objects.filter(tag=BACKEND_FIREBASE_CREDENTIALS).exists()) \
+            or (require_android and not FileAsText.objects.filter(tag=ANDROID_FIREBASE_CREDENTIALS).exists()) \
+            or (require_ios and not FileAsText.objects.filter(tag=IOS_FIREBASE_CREDENTIALS).exists()):
         return None
-    stored_credentials = json.loads(FileAsText.objects.get(tag=BACKEND_FIREBASE_CREDENTIALS).text)
-    return initialize_firebase_app(credentials.Certificate(stored_credentials))
-
-
-firebase_app = update_firebase_instance()
-
+    return get_firebase_app()
 
 class FirebaseNotCredentialed(Exception): pass
 class NoSchedulesException(Exception): pass
