+++ pages/admin_pages.py
@@ -1,19 +1,18 @@
+from django.core.exceptions import ValidationError
 from flask import (abort, Blueprint, flash, Markup, redirect, render_template, request,
     session)
-from wtforms import ValidationError
 
 from authentication import admin_authentication
 from authentication.admin_authentication import (authenticate_researcher_login,
                                                  authenticate_researcher_study_access, get_researcher_allowed_studies,
                                                  get_researcher_allowed_studies_as_query_set, get_session_researcher,
                                                  researcher_is_an_admin,
-                                                 SESSION_NAME, authenticate_admin)
+                                                 SESSION_NAME)
 from config.constants import BACKEND_FIREBASE_CREDENTIALS, ANDROID_FIREBASE_CREDENTIALS, IOS_FIREBASE_CREDENTIALS
 from config.settings import PUSH_NOTIFICATIONS_ENABLED
 from database.study_models import Study, StudyField
 from database.system_models import FileAsText
 from database.user_models import Participant, ParticipantFieldValue, Researcher
-from libs.push_notifications import update_firebase_instance
 from libs.security import check_password_requirements
 
 admin_pages = Blueprint('admin_pages', __name__)
@@ -183,107 +182,3 @@ def reset_download_api_credentials():
 def participant_tags(p: Participant):
         return {tag.field.field_name: tag.value for tag in p.field_values.all()}
 
-
-"""########################## FIREBASE CREDENTIALS ENDPOINTS ##################################"""
-
-
-@admin_pages.route('/manage_firebase_credentials')
-@authenticate_admin
-def manage_firebase_credentials():
-    return render_template('manage_firebase_credentials.html',
-                           firebase_credentials_exists=FileAsText.objects.filter(tag=BACKEND_FIREBASE_CREDENTIALS).exists(),
-                           #firebase_credentials_exists=False,
-                           android_credentials_exists=FileAsText.objects.filter(tag=ANDROID_FIREBASE_CREDENTIALS).exists(),
-                           ios_credentials_exists=FileAsText.objects.filter(tag=IOS_FIREBASE_CREDENTIALS).exists())
-
-
-@admin_pages.route('/upload_backend_firebase_cert', methods=['POST'])
-@authenticate_admin
-def upload_firebase_cert():
-    v = request.files['backend_firebase_cert']
-    try:
-        cert = str(v.read(), 'utf-8')
-        assert cert
-        FileAsText.objects.get_or_create(tag=BACKEND_FIREBASE_CREDENTIALS, defaults={"text": cert})
-        update_firebase_instance()
-        msg = """<h3>New firebase credentials have been received!</h3>"""
-        flash(Markup(msg), 'info')
-    except AssertionError:
-        msg = """
-                <div class="alert alert-danger" role="alert">
-                    <h3>There was an error in the processing the new firebase credentials!</h3>
-                    <p>You have selected no file or an empty file. If you just want to remove credentials, use the 
-                    delete button</p>
-                    <p>The previous credentials, if they existed, have not been removed</p>
-                </div>
-                """
-        flash(Markup(msg), 'error')
-    except (ValueError, UnicodeDecodeError, ValidationError) as error:
-        msg = f"""
-                <div class="alert alert-danger" role="alert">
-                    <h3>There was an error in the processing the new firebase credentials!</h3>
-                    <p>the error text read:
-                    {str(error)}</p>
-                </div>
-                """
-        flash(Markup(msg), 'error')
-    return redirect('/manage_firebase_credentials')
-
-
-@admin_pages.route('/upload_android_firebase_cert', methods=['POST'])
-@authenticate_admin
-def upload_android_firebase_cert():
-    v = request.files['android_firebase_cert']
-    cert = str(v.read(), 'utf-8')
-    FileAsText.objects.get_or_create(tag=ANDROID_FIREBASE_CREDENTIALS, defaults={"text": cert})
-    msg = """<h3>New android credentials were received!</h3>
-                <p>All registered android apps will be updated as they connect. That process may take some time</p>
-                """
-    flash(Markup(msg), 'info')
-    return redirect('/manage_firebase_credentials')
-
-
-@admin_pages.route('/upload_ios_firebase_cert', methods=['POST'])
-@authenticate_admin
-def upload_ios_firebase_cert():
-    v = request.files['ios_firebase_cert']
-    cert = str(v.read(), 'utf-8')
-    FileAsText.objects.get_or_create(tag=IOS_FIREBASE_CREDENTIALS, defaults={"text": cert})
-    msg = """<h3>New IOS credentials were received!</h3>
-                    <p>All registered IOS apps will be updated as they connect. That process may take some time</p>
-                    """
-    flash(Markup(msg), 'info')
-    return redirect('/manage_firebase_credentials')
-
-
-@admin_pages.route('/delete_backend_firebase_cert', methods=['POST'])
-@authenticate_admin
-def delete_backend_firebase_cert():
-    FileAsText.objects.get(tag=BACKEND_FIREBASE_CREDENTIALS).delete()
-    msg = """<h3>All backend Firebase credentials have been deleted!</h3>
-            <p>Note that this does not include IOS and Android app credentials, these must be deleted separately if 
-            desired</p>"""
-    flash(Markup(msg), 'info')
-    return redirect('/manage_firebase_credentials')
-
-
-@admin_pages.route('/delete_android_firebase_cert', methods=['POST'])
-@authenticate_admin
-def delete_android_firebase_cert():
-    FileAsText.objects.get(tag=ANDROID_FIREBASE_CREDENTIALS).delete()
-    msg = """<h3>Stored Android Firebase credentials have been removed if they existed!</h3>
-                <p>All registered android apps will be updated as they connect. That process may take some time</p>
-                """
-    flash(Markup(msg), 'info')
-    return redirect('/manage_firebase_credentials')
-
-
-@admin_pages.route('/delete_ios_firebase_cert', methods=['POST'])
-@authenticate_admin
-def delete_ios_firebase_cert():
-    FileAsText.objects.get(tag=IOS_FIREBASE_CREDENTIALS).delete()
-    msg = """<h3>Stored IOS Firebase credentials have been removed if they existed!</h3>
-                <p>All registered IOS apps will be updated as they connect. That process may take some time</p>
-                """
-    flash(Markup(msg), 'info')
-    return redirect('/manage_firebase_credentials')
